# Improve performance of Collection.removeFirst(_:) where Self == SubSequence