--- a/stdlib/public/core/Collection.swift
+++ b/stdlib/public/core/Collection.swift
@@ -1664,7 +1664,7 @@ extension Collection where SubSequence == Self {
   public mutating func removeFirst(_ k: Int) {
     if k == 0 { return }
     _precondition(k >= 0, "Number of elements to remove should be non-negative")
-    _precondition(count >= k,
+    _precondition(index(startIndex, offsetBy: k, limitedBy: endIndex) != nil,
       "Can't remove more items from a collection than it contains")
     self = self[index(startIndex, offsetBy: k)..<endIndex]
   }