@xwu 2020-06-18
@swift-ci Please smoke test