@theblixguy 2020-06-18
Windows failure is unrelated (needs https://github.com/apple/swift/pull/32454 & https://github.com/apple/swift/pull/32457 to unblock)

@swift-ci please smoke test Linux