@xwu 2020-06-18
I wonder, @stephencelis, would it be possible for you to incorporate this benchmark into the test suite in another PR? Then we can merge that first to show the nice shiny improvement with this PR while also ensuring no regressions on other benchmarks.