@theblixguy 2020-06-18
Here's a doc you might find handy: https://github.com/apple/swift/blob/master/benchmark/README.md.