@xwu 2020-06-18
There are a few files in the `single-source` directory that seem to be good examples. For instance, [ReduceInto.swift](https://github.com/apple/swift/blob/master/benchmark/single-source/ReduceInto.swift) seems pretty straightforward and might be a good starting point.