@natecook1000 2020-06-19
@stephencelis Thanks so much for identifying this issue and working on a fix! There are two more overloads in RangeReplaceableCollection.swift that could use this treatment, the second of which looks like the one that's actually getting used for `Substring`, since it's both range-replaceable and self-slicing. Could you look at those as well?