@stephencelis 2020-06-19
@natecook1000 Done!

@xwu Unfortunately I'm having issues getting things to run locally and I'm not sure I have a ton of free time to diagnose, so I'm afraid I won't be able to submit those benchmarks 😞