@natecook1000 2020-06-25
@swift-ci Please benchmark