@stephencelis 2020-06-25
Strange that SubstringDropFirst1 doesn't show up. Any idea why?