@natecook1000 2020-06-25
It's likely because I wrote that benchmark to test `dropFirst` instead of `removeFirst`… 🤦