@natecook1000 2020-06-26
@swift-ci Please benchmark